# docs/swagger.yaml
basePath: /
definitions:
...
  doc.EditProfile:
    properties:
      body:
        $ref: '#/definitions/dto.EditProfile'
    type: object
...
  doc.Profile:
    properties:
      body:
        $ref: '#/definitions/dto.Profile'
    type: object
  doc.ProfileArray:
    properties:
      body:
        items:
          $ref: '#/definitions/dto.Profile'
        type: array
    type: object
...
  doc.SignIn:
    properties:
      body:
        $ref: '#/definitions/dto.SignIn'
    type: object
  doc.SignUp:
    properties:
      body:
        $ref: '#/definitions/dto.SignUp'
    type: object
  doc.Subscribes:
    properties:
      body:
        $ref: '#/definitions/dto.Subscribes'
    type: object
...
  dto.EditProfile:
    properties:
      bio:
        type: string
      birthday:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      prev_pass:
        type: string
      sex:
        type: string
      status:
        type: string
      user_link:
        type: string
    type: object
...
  dto.Profile:
    properties:
      avatar:
        example: ""
        type: string
      bio:
        example: Текст с информацией о себе.
        type: string
      birthday:
        example: 30.04.2002
        type: string
      date_joined:
        example: 10.02.2023
        type: string
      first_name:
        example: Василий
        type: string
      last_active:
        example: ""
        type: string
      last_name:
        example: Петров
        type: string
      private:
        example: false
        type: boolean
      sex:
        example: male
        type: string
      status:
        example: Текст статуса.
        type: string
      user_link:
        example: id100500
        type: string
    type: object
  dto.SignIn:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.SignUp:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
  dto.Subscribes:
    properties:
      user_link:
        type: string
    type: object
...
  middleware.ErrorResponse:
    properties:
      message:
        example: Невалидный запрос.
        type: string
      status:
        example: 400
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: CashMap team
  description: Api for Depeche social network. VK Education project. Spring, 2023
  title: Depeche API
  version: "1.0"
paths:
...
  /api/user/friends:
    get:
      description: Get friends
      parameters:
      - description: link to requested profile
        in: query
        name: link
        required: true
        type: string
      - description: amount of profiles
        in: query
        name: limit
        required: true
        type: integer
      - description: number of batch
        in: query
        name: offset
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.ProfileArray'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Friends
      tags:
      - Profiles
  /api/user/profile:
    get:
      description: Get self profile
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Self
      tags:
      - Profiles
  /api/user/profile/edit:
    patch:
      consumes:
      - application/json
      description: Edit profile
      parameters:
      - description: Edited fields
        in: body
        name: request
        schema:
          $ref: '#/definitions/doc.EditProfile'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: EditProfile
      tags:
      - Profiles
  /api/user/profile/link:
    get:
      description: Get profile by link
      parameters:
      - description: link to requested profile
        in: path
        name: link
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Profile
      tags:
      - Profiles
  /api/user/reject:
    post:
      description: Reject friend request
      parameters:
      - description: Link to user to reject friend request.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/doc.Subscribes'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Reject
      tags:
      - Subscribes
  /api/user/sub:
    get:
      description: Get subscribes or subscribers for requested user
      parameters:
      - description: in/out for subscribers/subscribes
        in: query
        name: type
        required: true
        type: string
      - description: link to requested profile
        in: query
        name: link
        required: true
        type: string
      - description: amount of profiles
        in: query
        name: limit
        required: true
        type: integer
      - description: number of batch
        in: query
        name: offset
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/doc.ProfileArray'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Subscribes
      tags:
      - Profiles
    post:
      description: Subscribe to other user
      parameters:
      - description: Link to user to subscribe.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/doc.Subscribes'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Subscribe
      tags:
      - Subscribes
  /api/user/unsub:
    post:
      description: Unsubscribe from other user
      parameters:
      - description: Link to user to unsubscribe.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/doc.Subscribes'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Unsubscribe
      tags:
      - Subscribes
  /auth/logout:
    post:
      description: Delete user session and invalidate session cookie
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Log out
      tags:
      - Auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Authorize client with credentials (login and password).
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/doc.SignIn'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Sign in
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Register client with credentials and other user info.
      parameters:
      - description: Required register fields
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/doc.SignUp'
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Sign up
      tags:
      - Auth
swagger: "2.0"